test fund from funder

+++ FoundryBank/src/priceConverter.sol
@@ -5,17 +5,15 @@ pragma solidity ^0.8.16;
 import {AggregatorV3Interface} from "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
 library PriceConverter {
-
-    function getPrice(AggregatorV3Interface priceFeed) internal view returns (uint) {
+    function getPrice(AggregatorV3Interface priceFeed) internal view returns (uint256) {
         //abi chainlink registry contracto to see actual price
-        (,int256 price,,,) = priceFeed.latestRoundData();
+        (, int256 price,,,) = priceFeed.latestRoundData();
         return uint256(price * 1e10);
     }
 
-    function getConversionRate(uint ethAmount, AggregatorV3Interface priceFeed) internal view returns (uint) {
-        uint ethPrice = getPrice(priceFeed);
-        uint ethAmountInUSD = (ethPrice * ethAmount) / 1e18;
+    function getConversionRate(uint256 ethAmount, AggregatorV3Interface priceFeed) internal view returns (uint256) {
+        uint256 ethPrice = getPrice(priceFeed);
+        uint256 ethAmountInUSD = (ethPrice * ethAmount) / 1e18;
         return ethAmountInUSD;
     }
 }
-
